调整代码，同时解决rust analyzer未能提示warning的问题
