+++ kernel/src/arch/riscv64/riscv64gc-unknown-none-elf.json
@@ -18,7 +18,7 @@
     "std": false,
     "tier": 2
   },
-  "panic-strategy": "unwind",
+  "panic-strategy": "abort",
   "relocation-model": "static",
   "supported-sanitizers": [
     "shadow-call-stack",
