+++ kernel/src/debug/mod.rs
@@ -1,3 +1,4 @@
 pub mod jump_label;
 pub mod klog;
 pub mod kprobe;
+pub mod panic;
